pyproject.toml
[tool.poetry]
name = "bplus-rag-agent"
version = "0.1.0"
description = "BiasharaPlus Customer Support AI Agent"
authors = ["Samwel Ngusa"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.124.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
langchain = "^0.2.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.2.0"
langchain-pinecone = "^0.1.0"
pinecone-client = "^5.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.4.0"
tiktoken = "^0.7.0"
httpx = "^0.27.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"
pypdf = "^4.0.0"
unstructured = "^0.12.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
